and do NOT use a class for bootstrapping ?!?!
